Chris -> Jay:  I've added some commands to the console that will allow us to manipulate the scene a bit.
Commands implemented:
- loadscene -file <filename> -dir <directory>
- loadentity -file <filename> -dir <directory>
- loadplayervehicle -file <filename> -dir <directory>
- clearscene

In order to get anything to draw I need you to fill in the LoadMesh function in entity.cpp, and then create an inherited renderable class from entity and start calling your draw functions.

Also, gamestatemanager has been added.  The name is deceiving, however, because it is really just a wrapper for the scene.  Actual game state management occurs in appstate and bosapp.