Chris -> Jay:  I know that in your email you said that the problem has to do with the scaling of the models, and that's why they won't draw properly.  Today I tested out that code with a properly scaled model, and still nothing was working right.  Are you sure that the view frustum parameters are actually being set properly?
I was looking in our old code, and you have a whole bunch of extra stuff to setup the camers in game.cpp, in the Initialize() function.  I think that you might want to check to make sure that the view parameters are being set properly.

I copied the code that you had in game.cpp, and put it in gamestatemanager.cpp, in the initialize() function.  It's commented out right now, but I thought you might want to take a look at that stuff.